Make possible to combine scripts
Place code for scripts in ordinary classes so that they may be combined in scripts.
